requirements.txt
azure-mgmt-resource
azure-mgmt-storage
azure-cli-core
azure-mgmt-databricks